third_party/WebKit/LayoutTests/media/video-buffered-expected.txt

EXPECTED (video.buffered.length == '0') OK
TEST(video.buffered.start(-1)) THROWS(DOMException.INDEX_SIZE_ERR: Failed to execute 'start' on 'TimeRanges': The index provided (4294967295) is greater than the maximum bound (0).) OK
TEST(video.buffered.end(-1)) THROWS(DOMException.INDEX_SIZE_ERR: Failed to execute 'end' on 'TimeRanges': The index provided (4294967295) is greater than the maximum bound (0).) OK
TEST(video.buffered.start(0)) THROWS(DOMException.INDEX_SIZE_ERR: Failed to execute 'start' on 'TimeRanges': The index provided (0) is greater than or equal to the maximum bound (0).) OK
TEST(video.buffered.end(0)) THROWS(DOMException.INDEX_SIZE_ERR: Failed to execute 'end' on 'TimeRanges': The index provided (0) is greater than or equal to the maximum bound (0).) OK
EVENT(canplaythrough)
EXPECTED (video.buffered.length == '1') OK
EXPECTED (video.buffered.start(0) == '0') OK
EXPECTED (video.buffered.end(0) > '0') OK
TEST(video.buffered.start(-1)) THROWS(DOMException.INDEX_SIZE_ERR: Failed to execute 'start' on 'TimeRanges': The index provided (4294967295) is greater than the maximum bound (1).) OK
TEST(video.buffered.end(-1)) THROWS(DOMException.INDEX_SIZE_ERR: Failed to execute 'end' on 'TimeRanges': The index provided (4294967295) is greater than the maximum bound (1).) OK
TEST(video.buffered.start(1)) THROWS(DOMException.INDEX_SIZE_ERR: Failed to execute 'start' on 'TimeRanges': The index provided (1) is greater than or equal to the maximum bound (1).) OK
TEST(video.buffered.end(1)) THROWS(DOMException.INDEX_SIZE_ERR: Failed to execute 'end' on 'TimeRanges': The index provided (1) is greater than or equal to the maximum bound (1).) OK
END OF TEST
